hsh, home sweet home

A testing data set, a list of poly, line, points. The relationships between the objects cover many interesting cases. There are multi-island objects with holes and other objects that fall inside those holes. All paths share vertices with a neighbour. There is topological intersection along a diagonal line where two objects meet and are joined by shared vertices that were inserted deliberately.

Details

hsh is the holey home in sp format and hsh_sf in sf format.